Refactor BTC dominance chart related code

+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/coin/CoinViewFactory.kt
@@ -33,7 +33,7 @@ data class ChartPointViewItem(
     val date: Long,
     val price: CurrencyValue,
     val volume: CurrencyValue?,
-    val dominance: Float?,
+    val dominance: BigDecimal?,
     val macdInfo: MacdInfo?
 )
