+++ components/chartview/src/main/java/io/horizontalsystems/chartview/models/PointInfo.kt
@@ -1,9 +1,11 @@
 package io.horizontalsystems.chartview.models
 
+import java.math.BigDecimal
+
 class PointInfo(
     val value: Float,
     val volume: Float?,
     val macdInfo: MacdInfo?,
-    val dominance: Float?,
+    val dominance: BigDecimal?,
     val timestamp: Long
 )
